requirements.txt
boto==2.35.2
dj-database-url==0.3.0
Django==1.7.1
django-bootstrap-form==3.1
django-storages==1.1.1
flake8==2.3.0
mccabe==0.3
pep8==1.5.7
Pillow==2.7.0
pyflakes==0.8.1
waitress==0.8.9
psycopg2==2.5.4
whitenoise==1.0.6
